requirements.txt
django
groq
fastapi
uvicorn
tensorflow==2.18.0
keras==3.5.0
numpy
pillow
python-multipart
requests
